# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "grtblog 后端接口",
        "title": "grtblog API v2",
        "contact": {},
        "version": "2.0.0"
    },
    "basePath": "/api/v2",
    "paths": {
...
        "/pages": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Page"
                ],
                "summary": "获取页面列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "启用状态",
                        "name": "enabled",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "内置页面",
                        "name": "builtin",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.PageListResp"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Page"
                ],
                "summary": "创建页面",
                "parameters": [
                    {
                        "description": "创建页面参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.CreatePageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.PageResp"
                        }
                    }
                }
            }
        },
        "/pages/short/{shortUrl}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Page"
                ],
                "summary": "根据短链接获取页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短链接",
                        "name": "shortUrl",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.PageResp"
                        }
                    }
                }
            }
        },
        "/pages/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Page"
                ],
                "summary": "获取页面详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页面ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.PageResp"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Page"
                ],
                "summary": "更新页面",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页面ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新页面参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.UpdatePageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.PageResp"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Page"
                ],
                "summary": "删除页面",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页面ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {}
                    }
                }
            }
        },
        "/pages/{id}/latest": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Page"
                ],
                "summary": "校验页面是否最新",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页面ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "页面版本校验参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.CheckPageLatestReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.CheckPageLatestResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "contract.CheckPageLatestReq": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string"
                }
            }
        },
        "contract.CheckPageLatestResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "contentHash": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "latest": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "toc": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.TOCNode"
                    }
                }
            }
        },
...
        "contract.CreatePageReq": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "allowComment": {
                    "type": "boolean"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "extInfo": {
                    "type": "object"
                },
                "isBuiltin": {
                    "type": "boolean"
                },
                "isEnabled": {
                    "type": "boolean"
                },
                "shortUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "contract.MetricsResp": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
...
        "contract.PageListItemResp": {
            "type": "object",
            "properties": {
                "allowComment": {
                    "type": "boolean"
                },
                "commentAreaId": {
                    "type": "integer"
                },
                "comments": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isBuiltin": {
                    "type": "boolean"
                },
                "isEnabled": {
                    "type": "boolean"
                },
                "isHot": {
                    "type": "boolean"
                },
                "likes": {
                    "type": "integer"
                },
                "shortUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "contract.PageListResp": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.PageListItemResp"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "contract.PageResp": {
            "type": "object",
            "properties": {
                "aiSummary": {
                    "type": "string"
                },
                "allowComment": {
                    "type": "boolean"
                },
                "commentAreaId": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "contentHash": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "extInfo": {
                    "type": "object"
                },
                "id": {
                    "type": "integer"
                },
                "isBuiltin": {
                    "type": "boolean"
                },
                "isEnabled": {
                    "type": "boolean"
                },
                "isHot": {
                    "type": "boolean"
                },
                "metrics": {
                    "$ref": "#/definitions/contract.MetricsResp"
                },
                "shortUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "toc": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.TOCNode"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "contract.TOCNode": {
            "type": "object",
            "properties": {
                "anchor": {
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contract.TOCNode"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "contract.UpdatePageReq": {
            "type": "object",
            "required": [
                "content",
                "shortUrl",
                "title"
            ],
            "properties": {
                "allowComment": {
                    "type": "boolean"
                },
                "content": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "extInfo": {
                    "type": "object"
                },
                "isBuiltin": {
                    "type": "boolean"
                },
                "isEnabled": {
                    "type": "boolean"
                },
                "shortUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}